# docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Holy Hymns API",
    "version": "1.0.0",
    "description": "Self-hosted lyrics catalogue. Session authorization is checked on every protected request. All timestamps UTC. Public responses never contain drafts."
  },
  "servers": [
    {
      "url": "http://localhost:8080/v1",
      "description": "Local development"
    }
  ],
  "paths": {
    "/songs": {
      "get": {
        "summary": "Published songs, ranked bilingual search",
        "security": [],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SongList"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Missing, expired or revoked session"
          },
          "403": {
            "description": "Role or origin denied"
          },
          "404": {
            "description": "Resource unavailable"
          },
          "409": {
            "description": "Version or identity conflict"
          },
          "429": {
            "description": "Rate or free email quota exhausted"
          },
          "503": {
            "description": "Dependency unavailable"
          },
          "500": {
            "description": "Unexpected server or database failure",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "category",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "letter",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "featured",
            "in": "query",
            "schema": {
              "type": "boolean",
              "default": false
            }
          },
          {
            "name": "sort",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "title",
                "recent"
              ],
              "default": "title"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 30,
              "maximum": 100
            }
          },
          {
            "name": "offset",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "maximum": 100000
            }
          }
        ]
      }
    },
...
    "/auth/login": {
      "post": {
        "summary": "Sign in with verified email and password",
        "security": [],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Session"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Missing, expired or revoked session"
          },
          "403": {
            "description": "Role or origin denied"
          },
          "404": {
            "description": "Resource unavailable"
          },
          "409": {
            "description": "Version or identity conflict"
          },
          "429": {
            "description": "Rate or free email quota exhausted"
          },
          "503": {
            "description": "Dependency unavailable"
          },
          "500": {
            "description": "Unexpected server or database failure",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": {
                    "type": "string"
                  },
                  "password": {
                    "type": "string"
                  }
                },
                "required": [
                  "email",
                  "password"
                ],
                "additionalProperties": false
              }
            }
          }
        },
        "description": "Only verified, active accounts may sign in. Response token is an opaque revocable session. Attempts are limited by trusted client IP and account."
      }
    },
...
    "/auth/logout": {
      "post": {
        "summary": "Revoke current session",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Message"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Missing, expired or revoked session"
          },
          "403": {
            "description": "Role or origin denied"
          },
          "404": {
            "description": "Resource unavailable"
          },
          "409": {
            "description": "Version or identity conflict"
          },
          "429": {
            "description": "Rate or free email quota exhausted"
          },
          "503": {
            "description": "Dependency unavailable"
          },
          "500": {
            "description": "Unexpected server or database failure",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
    "/admin/imports/lyrics": {
      "post": {
        "summary": "Import one extracted Grejo Lyrics post as a private draft",
        "operationId": "importExtractedLyrics",
        "description": "Administrator or owner only. Uses durable source ID/hash deduplication and never overwrites app edits or publishes content. Matching source/hash returns unchanged; a changed hash returns conflicts in the 200 report. A non-dry import commits draft, taxonomy, first revision, import run and lyrics.import actor audit atomically. Retry failed requests safely. Existing 1 MiB body limit applies; only one JSON object and documented fields are accepted.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LyricsImportRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Import result or dry-run preview",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImportReport"
                }
              }
            }
          },
          "400": {
            "description": "Invalid source, metadata, song content, JSON or oversized request body",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Missing, expired or revoked session"
          },
          "403": {
            "description": "Reader or origin denied"
          },
          "500": {
            "description": "Import or audit failed; retry safely using the same sourceId and hash. No partial import can commit.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "503": {
            "description": "Server busy or transaction timed out; retry safely"
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "Opaque revocable account session, not a provider identity token."
      }
    },
    "schemas": {
      "Song": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "title": {
            "type": "string"
          },
          "titleMalayalam": {
            "type": "string"
          },
          "lyricsMalayalam": {
            "type": "string"
          },
          "lyricsManglish": {
            "type": "string"
          },
          "aliases": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "credits": {
            "type": "string"
          },
          "links": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/MediaLink"
            }
          },
          "categoryIds": {
            "type": "array",
            "items": {
              "type": "string",
              "format": "uuid"
            }
          },
          "featured": {
            "type": "boolean"
          },
          "version": {
            "type": "integer"
          },
          "status": {
            "type": "string",
            "enum": [
              "draft",
              "published"
            ]
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "sourceUrl": {
            "type": "string"
          },
          "reviewNotes": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Administrative review notes; omitted from public song responses."
          }
        },
        "required": [
          "id",
          "title",
          "titleMalayalam",
          "lyricsMalayalam",
          "lyricsManglish",
          "aliases",
          "credits",
          "links",
          "categoryIds",
          "featured",
          "version",
          "status",
          "updatedAt"
        ]
      },
      "User": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "email": {
            "type": "string",
            "format": "email"
          },
          "name": {
            "type": "string"
          },
          "role": {
            "type": "string",
            "enum": [
              "reader",
              "admin",
              "owner"
            ]
          },
          "verified": {
            "type": "boolean"
          },
          "suspended": {
            "type": "boolean"
          }
        },
        "required": [
          "id",
          "email",
          "name",
          "role",
          "verified",
          "suspended"
        ]
      },
...
      "MediaLink": {
        "type": "object",
        "properties": {
          "label": {
            "type": "string"
          },
          "url": {
            "type": "string",
            "format": "uri",
            "description": "An absolute HTTPS URL without user-info; no media is fetched or hosted by the API."
          }
        },
        "required": [
          "url"
        ]
      },
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string"
          },
          "code": {
            "type": "string"
          }
        },
        "required": [
          "error"
        ]
      },
      "Message": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string"
          }
        },
        "required": [
          "message"
        ]
      },
      "Session": {
        "type": "object",
        "properties": {
          "token": {
            "type": "string"
          },
          "user": {
            "$ref": "#/components/schemas/User"
          }
        },
        "required": [
          "token",
          "user"
        ]
      },
      "SongList": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Song"
            }
          },
          "total": {
            "type": "integer"
          }
        },
        "required": [
          "items",
          "total"
        ]
      },
...
      "LyricsImportEntry": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "sourceId",
          "sourceUrl",
          "hash"
        ],
        "description": "One extracted Grejo Lyrics post. At least one nonblank title is required. Lyrics together are limited to 200,000 UTF-8 bytes; no usable lyrics returns a skipped report. Null characters and RawHTML are not accepted. Labels are trimmed and deduplicated. Hash is normalized to lowercase.",
        "properties": {
          "sourceId": {
            "type": "string",
            "pattern": "^tag:blogger\\.com,1999:blog-6936100729546217639\\.post-[0-9]{1,32}$"
          },
          "sourceUrl": {
            "type": "string",
            "format": "uri",
            "maxLength": 2048,
            "pattern": "^https://grejolyrics\\.blogspot\\.com/[0-9]{4}/(0[1-9]|1[0-2])/[A-Za-z0-9_-][A-Za-z0-9._~-]*\\.html$",
            "description": "Canonical HTTPS post permalink only; maximum 2048 UTF-8 bytes. No credentials, port, query, fragment or escaped path."
          },
          "hash": {
            "type": "string",
            "pattern": "^[a-fA-F0-9]{64}$"
          },
          "title": {
            "type": "string",
            "maxLength": 500
          },
          "titleMalayalam": {
            "type": "string",
            "maxLength": 500
          },
          "lyricsMalayalam": {
            "type": "string"
          },
          "lyricsManglish": {
            "type": "string"
          },
          "credits": {
            "type": "string",
            "maxLength": 4000
          },
          "labels": {
            "type": "array",
            "maxItems": 30,
            "items": {
              "type": "string",
              "minLength": 1,
              "maxLength": 120,
              "description": "Nonblank after trimming; no line breaks."
            }
          },
          "reviewNotes": {
            "type": "array",
            "maxItems": 30,
            "items": {
              "type": "string",
              "minLength": 1,
              "maxLength": 1000
            }
          },
          "links": {
            "type": "array",
            "maxItems": 10,
            "items": {
              "type": "object",
              "additionalProperties": false,
              "required": [
                "label",
                "url"
              ],
              "properties": {
                "label": {
                  "type": "string",
                  "minLength": 1,
                  "maxLength": 120
                },
                "url": {
                  "type": "string",
                  "format": "uri",
                  "maxLength": 2048,
                  "pattern": "^https://",
                  "description": "HTTPS URL with host and without credentials; at most 2048 UTF-8 bytes."
                }
              }
            }
          }
        },
        "anyOf": [
          {
            "required": [
              "title"
            ],
            "properties": {
              "title": {
                "minLength": 1
              }
            }
          },
          {
            "required": [
              "titleMalayalam"
            ],
            "properties": {
              "titleMalayalam": {
                "minLength": 1
              }
            }
          }
        ]
      },
      "LyricsImportRequest": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "entry"
        ],
        "properties": {
          "entry": {
            "$ref": "#/components/schemas/LyricsImportEntry"
          },
          "dryRun": {
            "type": "boolean",
            "default": false,
            "description": "Read-only validation and preview; no song, category, revision, import run or audit is written."
          }
        }
      },
      "ImportReport": {
        "type": "object",
        "required": [
          "total",
          "created",
          "unchanged",
          "conflicts",
          "skipped",
          "review"
        ],
        "properties": {
          "total": {
            "type": "integer",
            "minimum": 0
          },
          "created": {
            "type": "integer",
            "minimum": 0
          },
          "unchanged": {
            "type": "integer",
            "minimum": 0
          },
          "conflicts": {
            "type": "integer",
            "minimum": 0
          },
          "skipped": {
            "type": "integer",
            "minimum": 0
          },
          "review": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      }
    }
  }
}